requirements.txt
# requirements.txt
streamlit
streamlit-tree-select
PyYAML
langchain-openai
langchain-community
langchain
requests